+++ R/formatdown-package.R
@@ -1,7 +1,9 @@
 
-#' @importFrom checkmate qassert assert_disjunct assert_choice
+#' @importFrom checkmate qassert assert_disjunct assert_choice assert_true
 #'
-#' @importFrom data.table copy setDT fcase :=  %like% %between%
+#' @importFrom data.table copy setDT fifelse fcase :=  %like% %between%
+#'
+#' @importFrom units as_units deparse_unit
 #'
 #' @importFrom wrapr stop_if_dot_args %?%
 NULL