+++ man/examples/examples_format_power.R
@@ -1,5 +1,5 @@
 # Scalar value
-format_power(101100)
+format_power(101100, digits = 4)
 
 # Vector value
 x <- c(1.2222e-6, 2.3333e-5, 3.4444e-4, 4.1111e-3, 5.2222e-2, 6.3333e-1,
@@ -13,6 +13,11 @@ format_power(x[1], 4)
 # Compare format type
 format_power(x[3], format = "engr")
 format_power(x[3], format = "sci")
+
+# Compare set_power results
+format_power(x[3], set_power = -5)
+format_power(x[3], set_power = -4)
+format_power(x[3], set_power = -3)
 
 # Compare omit_power range
 format_power(x[6], omit_power = c(-1, 2))